requirements.txt
flask==3.0.0
selenium==4.16.0
python-dotenv==1.0.0
flask-cors==4.0.0
chromedriver-autoinstaller==0.6.4
